requirements.txt
requests==2.32.3
pydantic==2.7.4
psycopg2-binary==2.9.9

